ConsumerRecords are decrypted.
Process flow incorporates a call to DKS to decrypt the datakey. The
encrypted dbObject is then decrypted with the decrypted datakey.

At present decrypted records are simply posted to a topic with a name
derived from the source record topic.

Records that fail processing are sent to the dead letter queue.