All db connection parameters fetched from secrets manager.
It was assumed that only the database password was in secrets manager, however all
connection parameters are there in a JSON structure so fetch them all rather
than having them as individual properties (the redundant configuration parameters
have been removed).

Add the RDS CA cert and use it by default when establishing a database connection.